requirements.txt
torch_cluster
tqdm
opencv-python
scipy
ipdb
matplotlib
cutoop
tensorboardX
hydra-core
scikit-learn